Initial commit (initialize the workspace NOW)
Before continuing, you must initialize the workspace. Instructions are
provided here:
These archives (which have identical contents, and are provided for
convenience) contain initial metadata that is necessary to start using
this repository. The files themselves are not tracked in the normal
repository for various reasons, so this one-time process should allow
you to use eclipse with this repository efficiently.