Restructured Drive components
Code now uses encapsulation properly, i.e. the DriveBase class
doesn't need to know that it's a part of a larger robot to work.

This means dropping the lookalike functions from the Robot class,
because the DriveBase shouldn't have to know that it's in
autonamous mode, or periodic mode; the Robot class should handle
that, and tell its components what to do based on what it knows.