Fixed port numbers and joystick values
The port numbers for the drive train were reversed, but the joystick
values weren't being inverted when they should've. This resulted in
the two problems cancelling each other out. However, this was a
problem while debugging, and is now fixed. Also, the left side is now
boosted when going forwards so it'll drive straight.

This could have been found sooner if we'd used an interface for
setting the motor speed instead of passing a pointer to the xbox
object to everything that needs to get input.